requirements.txt
Django>=3.0,<4.0
djangorestframework>=3.12.4
djangorestframework-simplejwt==5.2.2
